atof changes

+++ ft_pow.c
@@ -6,15 +6,15 @@
 /*   By: wkornato <wkornato@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 10:46:26 by wkornato          #+#    #+#             */
-/*   Updated: 2024/08/20 11:05:08 by wkornato         ###   ########.fr       */
+/*   Updated: 2024/08/20 11:26:47 by wkornato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_pow(double x, int y)
+long double	ft_pow(double x, int y)
 {
-	double	res;
+	long double	res;
 
 	res = 1;
 	while (y > 0)
